Remove unused imports extract CafeModel
